allow args compatible with browserify
this change adds another way to pass arguments into `cuke-tap`,
in a way that's compatible with being bundled with `browserify`.
basically, `gherkin` needs the feature file content and the feature
file path, so the new argument signature is passing in an array
of arrays with this information.

also use latest version of gherkin on GitHub.